+++ Configuration/TCA/tx_realty_pets.php
@@ -13,7 +13,7 @@ return [
         'transOrigDiffSourceField' => 'l18n_diffsource',
         'default_sortby' => 'ORDER BY title',
         'delete' => 'deleted',
-        'iconfile' => 'EXT:realty/icons/icon_tx_realty_pets.gif',
+        'iconfile' => 'EXT:realty/Resources/Public/Icons/Pet.gif',
         'searchFields' => 'uid,title',
     ],
     'interface' => [