+++ ext_emconf.php
@@ -3,19 +3,8 @@
 $EM_CONF[$_EXTKEY] = [
     'title' => 'Realty Manager',
     'description' => 'Provides a plugin that displays realty objects (properties, real estate), including an image gallery for each object.',
-    'category' => 'plugin',
-    'author' => 'Oliver Klee',
-    'author_email' => 'typo3-coding@oliverklee.de',
-    'shy' => 0,
-    'dependencies' => 'oelib,mkforms,static_info_tables',
-    'conflicts' => 'dbal',
-    'state' => 'stable',
-    'createDirs' => 'uploads/tx_realty/,uploads/tx_realty/rte/',
-    'modify_tables' => 'fe_users',
-    'clearCacheOnLoad' => 1,
-    'author_company' => 'oliverklee.de',
     'version' => '1.1.0',
-    '_md5_values_when_last_written' => '',
+    'category' => 'plugin',
     'constraints' => [
         'depends' => [
             'php' => '5.5.0-7.2.99',
@@ -33,6 +22,13 @@ $EM_CONF[$_EXTKEY] = [
             'sr_feuser_register' => '',
         ],
     ],
+    'state' => 'stable',
+    'uploadfolder' => true,
+    'createDirs' => 'uploads/tx_realty/,uploads/tx_realty/rte/',
+    'clearCacheOnLoad' => true,
+    'author' => 'Oliver Klee',
+    'author_email' => 'typo3-coding@oliverklee.de',
+    'author_company' => 'oliverklee.de',
     'autoload' => [
         'classmap' => [
             'Classes',